// src/tcgdb/modules/tcgdb/api/mtg/price/change/weekly/dto/tcgdb.mtg.price.change.weekly.dto.ts
export class TCGdbMTGPricesChangeWeeklyDTO {
    tcgdbMTGPricesChangeWeekly: TCGdbMTGPriceChangeWeeklyDTO[];
}

export class TCGdbMTGPriceChangeWeeklyDTO {
    tcgdbMTGPriceChangeWeeklyId: string;
    tcgdbMTGCardId: string;
    tcgdbMTGPriceChangeWeeklyTCGPlayerId: number;
    tcgdbMTGPriceChangeWeeklySetCode: string;
    tcgdbMTGPriceChangeWeeklyCurrentLowPrice: number;
    tcgdbMTGPriceChangeWeeklyPreviousLowPrice: number;
    tcgdbMTGPriceChangeWeeklyLowPriceDifference: number;
    tcgdbMTGPriceChangeWeeklyLowPricePercentage: number;
    tcgdbMTGPriceChangeWeeklyCurrentMidPrice: number;
    tcgdbMTGPriceChangeWeeklyPreviousMidPrice: number;
    tcgdbMTGPriceChangeWeeklyMidPriceDifference: number;
    tcgdbMTGPriceChangeWeeklyMidPricePercentage: number;
    tcgdbMTGPriceChangeWeeklyCurrentHighPrice: number;
    tcgdbMTGPriceChangeWeeklyPreviousHighPrice: number;
    tcgdbMTGPriceChangeWeeklyHighPriceDifference: number;
    tcgdbMTGPriceChangeWeeklyHighPricePercentage: number;
    tcgdbMTGPriceChangeWeeklyCurrentMarketPrice: number;
    tcgdbMTGPriceChangeWeeklyPreviousMarketPrice: number;
    tcgdbMTGPriceChangeWeeklyMarketPriceDifference: number;
    tcgdbMTGPriceChangeWeeklyMarketPricePercentage: number;
    tcgdbMTGPriceChangeWeeklySubTypeName: string;
}

export class CreateTCGdbMTGPriceChangeWeeklyDTO {
    tcgdbMTGCardId: string;
    tcgdbMTGPriceChangeWeeklyTCGPlayerId: number;
    tcgdbMTGPriceChangeWeeklySetCode: string;
    tcgdbMTGPriceChangeWeeklyCurrentLowPrice: number;
    tcgdbMTGPriceChangeWeeklyPreviousLowPrice: number;
    tcgdbMTGPriceChangeWeeklyLowPriceDifference: number;
    tcgdbMTGPriceChangeWeeklyLowPricePercentage: number;
    tcgdbMTGPriceChangeWeeklyCurrentMidPrice: number;
    tcgdbMTGPriceChangeWeeklyPreviousMidPrice: number;
    tcgdbMTGPriceChangeWeeklyMidPriceDifference: number;
    tcgdbMTGPriceChangeWeeklyMidPricePercentage: number;
    tcgdbMTGPriceChangeWeeklyCurrentHighPrice: number;
    tcgdbMTGPriceChangeWeeklyPreviousHighPrice: number;
    tcgdbMTGPriceChangeWeeklyHighPriceDifference: number;
    tcgdbMTGPriceChangeWeeklyHighPricePercentage: number;
    tcgdbMTGPriceChangeWeeklyCurrentMarketPrice: number;
    tcgdbMTGPriceChangeWeeklyPreviousMarketPrice: number;
    tcgdbMTGPriceChangeWeeklyMarketPriceDifference: number;
    tcgdbMTGPriceChangeWeeklyMarketPricePercentage: number;
    tcgdbMTGPriceChangeWeeklySubTypeName: string;
}